+++ MeetupApp/MeetupApp/MeetupApp/ViewModels/MainPageViewModel.cs
@@ -1,25 +1,13 @@
 ﻿using Prism.Navigation;
-using System.Windows.Input;
-using MeetupApp.Commands;
-using MeetupApp.Services;
 
 namespace MeetupApp.ViewModels
 {
-	public class MainPageViewModel : ViewModelBase, INavigationAware
+    public class MainPageViewModel : ViewModelBase
     {
-        private ICommand _loadEventsCommand;
-        public ICommand RefreshCommand => _loadEventsCommand;
-
-        public MainPageViewModel(INavigationService navigationService, ILoadEventsCommand loadEventsCommand)
-            : base (navigationService)
+        public MainPageViewModel(INavigationService navigationService)
+            : base(navigationService)
         {
             Title = "Main Page";
-            _loadEventsCommand = loadEventsCommand;
         }
-
-		public override void OnNavigatingTo(NavigationParameters parameters)
-        {
-            _loadEventsCommand.Execute(null);
-        }
-	}
+    }
 }
