+++ MeetupApp/MeetupApp/MeetupApp/Views/AboutPage.xaml
@@ -3,8 +3,5 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml" 
              xmlns:local="clr-namespace:MeetupApp.Views"
              x:Class="MeetupApp.Views.AboutPage">
-    <local:BasePage.Resources>
-        <!-- Application resource dictionary -->
-        <StyleSheet Source="../Styles/core.css" />
-    </local:BasePage.Resources>
+    <Label Text="About"/>
 </local:BasePage>
