+++ MeetupApp/MeetupApp/MeetupApp/Views/MeetingsListPage.xaml
@@ -1,13 +1,8 @@
-<?xml version="1.0" encoding="UTF-8"?>
+<?xml version="1.0" encoding="utf-8" ?>
 <local:BasePage xmlns="http://xamarin.com/schemas/2014/forms"
-                xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml" 
-                xmlns:local="clr-namespace:MeetupApp.Views"
-                x:Class="MeetupApp.Views.MeetingsListPage">
-    <local:BasePage.Resources>
-
-        <!-- Application resource dictionary -->
-        <StyleSheet Source="../Styles/core.css" />
-    </local:BasePage.Resources>
+             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+             xmlns:local="clr-namespace:MeetupApp.Views"
+             x:Class="MeetupApp.Views.MeetingsListPage">
     <Grid ColumnSpacing="0" RowSpacing="0">
         <Grid.RowDefinitions>
             <RowDefinition Height="*"/>
@@ -31,10 +26,19 @@
             <Label Text="Toronto Mobile .NET Developers" class="h1"
                    VerticalOptions="CenterAndExpand"/>
         </StackLayout>
-        <ScrollView Grid.Row="3" Grid.ColumnSpan="3">
-            <StackLayout class="content">
-                <Label Text="stuff"/>
-            </StackLayout>
-        </ScrollView>
+        <ListView Grid.Row="3" Grid.ColumnSpan="3" class="content"
+                   ItemsSource="{Binding RefreshCommand.EventList}"
+                  IsPullToRefreshEnabled="True"
+                  RefreshCommand="{Binding RefreshCommand}"
+                  IsRefreshing="{Binding RefreshCommand.IsRefreshing}">
+            <ListView.ItemTemplate>
+                <DataTemplate>
+                    <!--<ViewCell>
+                        <Label Text="{Binding Title}"/>
+                    </ViewCell>-->
+                    <TextCell Text="{Binding Title}" /><!-- Detail="{Binding Description}" />-->
+                </DataTemplate>
+            </ListView.ItemTemplate>
+        </ListView>
     </Grid>
 </local:BasePage>
